graphql: Move validate_fields into execution::Query
